requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
et-xmlfile==1.1.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-vision==3.4.4
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==2.0.6
